Add popup help capabilities
Help icons can be added anywhere on RT pages. When clicked, they
will display the contents of an associated Article in a dismissable
popover dialog. The Article content can be retrieved server-side
at page rendering time or dynamically when clicked via AJAX.

No help icons or articles are actually added as part of this commit.
Just the underlying functionality.